Comments fixed
