+++ seedpod_ground_risk/ui_resources/add_layer_wizard.py
@@ -7,6 +7,7 @@ from PySide2.QtGui import QRegExpValidator
 from PySide2.QtWidgets import QWizard, QWizardPage, QLabel, QLineEdit, QComboBox, QCheckBox, QGridLayout, QColorDialog, \
     QPushButton, QFileDialog
 
+from seedpod_ground_risk.ui_resources.aircraft_options import AIRCRAFT_LIST
 from seedpod_ground_risk.ui_resources.coordbox import GeoWidget
 from seedpod_ground_risk.ui_resources.layer_options import *
 
@@ -145,7 +146,7 @@ class LayerWizard(QWizard):
         for name, opt in list(LAYER_OPTIONS.values())[self.layerType].items():
             if opt[1] == 'algo':
                 d = {opt[1]: list(ALGORITHM_OBJECTS.values())[self.field(name)]}
-            elif opt[1] == 'coordinate':
+            elif opt[0] == 'coordinate':
                 d = {opt[1]: (self.field(name).latitude(), self.field(name).longitude())}
             elif opt[1] == 'aircraft':
                 d = {opt[1]: list(AIRCRAFT_LIST.values())[self.field(name)]}
