+++ seedpod_ground_risk/ui_resources/layer_options.py
@@ -1,5 +1,3 @@
-import json
-
 from seedpod_ground_risk.layers.arbitrary_obstacle_layer import ArbitraryObstacleLayer
 from seedpod_ground_risk.layers.osm_tag_layer import OSMTagLayer
 from seedpod_ground_risk.layers.path_analysis_layer import PathAnalysisLayer
@@ -80,21 +78,3 @@ AIRCRAFT_PARAMETERS = {
     'Flight Airspeed [m/s]': (r'-?\d{0,3}\.?\d+', 'vel', float)
 }
 
-
-# Create aircraft list dictionary from UAV list found in static_data
-
-def aircraft_list(filepath="static_data/aircraft_list.json"):
-    json_file_path = filepath
-    with open(json_file_path, 'r') as j:
-        aircrafts = json.loads(j.read())
-    return aircrafts
-
-
-def add_aircraft(new_ac):
-    ac_list = AIRCRAFT_LIST
-    ac_list[f"{new_ac['name']}"] = new_ac
-    if 'name' in ac_list[f"{new_ac['name']}"]: del ac_list[f"{new_ac['name']}"]['name']
-    with open("static_data/aircraft_list.json", 'w') as f:
-        json.dump(ac_list, f)
-
-AIRCRAFT_LIST = aircraft_list()
